requirements.txt
boto3==1.35.54
botocore==1.35.54
certifi==2024.8.30
charset-normalizer==3.4.0
feedparser==6.0.11
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.3
sgmllib3k==1.0.0
six==1.16.0
urllib3==2.2.3
